mostly done with refactor
