+++ src/generation/mod.rs
@@ -1,1 +1,1 @@
-pub mod parsed_to_waveform;
+pub mod parsed_to_render;
