--- src/operations/apply_tests.rs
@@ -1,374 +1,0 @@
-pub mod apply_tests {
-    extern crate num_rational;
-    use event::{Event, Sound};
-    use num_rational::Ratio;
-    use operations::{Apply, GetLengthRatio};
-    use socool_parser::ast::Op;
-
-    fn event1() -> Event {
-        Event {
-            sounds: vec![Sound {
-                frequency: 100.0,
-                gain: 1.0,
-                pan: 0.0,
-            }],
-            length: 1.0,
-        }
-    }
-
-    fn event2() -> Event {
-        Event {
-            sounds: vec![Sound {
-                frequency: 100.0,
-                gain: 1.0,
-                pan: 0.0,
-            }],
-            length: 1.0,
-        }
-    }
-
-    fn vec_event1() -> Vec<Event> {
-        vec![event1(), event2()]
-    }
-
-    fn sequence1() -> Op {
-        Op::Sequence {
-            operations: vec![
-                Op::AsIs,
-                Op::AsIs,
-                Op::TransposeM {
-                    m: Ratio::new(2, 1),
-                },
-                Op::Length {
-                    m: Ratio::new(2, 1),
-                },
-            ],
-        }
-    }
-
-    fn sequence2() -> Op {
-        Op::Compose {
-            operations: vec![
-                sequence1(),
-                Op::Length {
-                    m: Ratio::new(2, 1),
-                },
-            ],
-        }
-    }
-
-    fn sequence3() -> Op {
-        Op::Compose {
-            operations: vec![sequence1(), sequence2()],
-        }
-    }
-
-    #[test]
-    fn op_asis_test() {
-        let as_is = Op::AsIs {};
-        assert_eq!(as_is.get_length_ratio(), Ratio::new(1, 1));
-        assert_eq!(as_is.apply(vec_event1()), vec_event1())
-    }
-
-    #[test]
-    fn op_reverse_test() {
-        let reverse = Op::Reverse {};
-        assert_eq!(reverse.get_length_ratio(), Ratio::new(1, 1));
-
-        let apply_expected = vec![event2(), event1()];
-        assert_eq!(reverse.apply(vec_event1()), apply_expected);
-    }
-
-    #[test]
-    fn op_pan_test() {
-        let pan_a = Op::PanA {
-            a: Ratio::new(1, 2),
-        };
-        let pan_m = Op::PanM {
-            m: Ratio::new(1, 2),
-        };
-        assert_eq!(pan_a.get_length_ratio(), Ratio::new(1, 1));
-        assert_eq!(pan_m.get_length_ratio(), Ratio::new(1, 1));
-
-        let mut expected_event = event1();
-        expected_event.sounds[0].pan = 0.5;
-        let apply_expected = vec![expected_event];
-        assert_eq!(pan_a.apply(vec![event1()]), apply_expected);
-
-        let mut expected_event = event1();
-        expected_event.sounds[0].pan = 0.0;
-        let apply_expected = vec![expected_event];
-        assert_eq!(pan_m.apply(vec![event1()]), apply_expected);
-    }
-
-    #[test]
-    fn op_length_test() {
-        let length = Op::Length {
-            m: Ratio::new(3, 2),
-        };
-        assert_eq!(length.get_length_ratio(), Ratio::new(3, 2));
-
-        let mut expected_event = event1();
-        expected_event.length = 1.5;
-        let apply_expected = vec![expected_event];
-        assert_eq!(length.apply(vec![event1()]), apply_expected);
-    }
-
-    #[test]
-    fn op_transpose_m_test() {
-        let transpose_m = Op::TransposeM {
-            m: Ratio::new(3, 2),
-        };
-        assert_eq!(transpose_m.get_length_ratio(), Ratio::new(1, 1));
-
-        let mut expected_event = event1();
-        expected_event.sounds[0].frequency = 150.0;
-        let apply_expected = vec![expected_event];
-        assert_eq!(transpose_m.apply(vec![event1()]), apply_expected);
-    }
-
-    #[test]
-    fn op_transpose_a_test() {
-        let transpose_a = Op::TransposeA {
-            a: Ratio::new(3, 2),
-        };
-        assert_eq!(transpose_a.get_length_ratio(), Ratio::new(1, 1));
-
-        let mut expected_event = event1();
-        expected_event.sounds[0].frequency = 101.5;
-        let apply_expected = vec![expected_event];
-        assert_eq!(transpose_a.apply(vec![event1()]), apply_expected);
-    }
-
-    #[test]
-    fn op_silence_test() {
-        let silence = Op::Silence {
-            m: Ratio::new(3, 2),
-        };
-        assert_eq!(silence.get_length_ratio(), Ratio::new(3, 2));
-
-        let mut expected_event = event1();
-        expected_event.sounds[0].frequency = 0.0;
-        expected_event.sounds[0].gain = 0.0;
-        expected_event.length = 1.5;
-        let apply_expected = vec![expected_event];
-        assert_eq!(silence.apply(vec![event1()]), apply_expected);
-    }
-
-    #[test]
-    fn op_gain_test() {
-        let gain = Op::Gain {
-            m: Ratio::new(3, 2),
-        };
-        assert_eq!(gain.get_length_ratio(), Ratio::new(1, 1));
-
-        let mut expected_event = event1();
-        expected_event.sounds[0].gain = 1.5;
-        let apply_expected = vec![expected_event];
-        assert_eq!(gain.apply(vec![event1()]), apply_expected);
-    }
-
-    #[test]
-    #[cfg_attr(rustfmt, rustfmt_skip)]
-    fn op_fit() {
-        let fit = Op::WithLengthRatioOf {
-            with_length_of: Box::new(sequence1()),
-            main: Box::new(sequence2()),
-        };
-
-        let fit_length = fit.get_length_ratio();
-        assert_eq!(fit_length, Ratio::new(5, 1));
-
-        let expected =
-            vec![Event { sounds: vec![Sound { frequency: 100.0, gain: 1.0, pan: 0.0 }], length: 0.5 },
-                 Event { sounds: vec![Sound { frequency: 100.0, gain: 1.0, pan: 0.0 }], length: 0.5 }];
-
-        assert_eq!(fit.apply(vec_event1()), expected);
-    }
-
-    #[test]
-    fn op_compose_test() {
-        let sequence_with_sequence_length = sequence3().get_length_ratio();
-        assert_eq!(sequence_with_sequence_length, Ratio::new(50, 1));
-
-        let compose = Op::Compose {
-            operations: vec![
-                Op::Length {
-                    m: Ratio::new(2, 1),
-                },
-                Op::TransposeM {
-                    m: Ratio::new(3, 2),
-                },
-            ],
-        };
-
-        let expected = vec![
-            Event {
-                sounds: vec![Sound {
-                    frequency: 150.0,
-                    gain: 1.0,
-                    pan: 0.0,
-                }],
-                length: 2.0,
-            },
-            Event {
-                sounds: vec![Sound {
-                    frequency: 150.0,
-                    gain: 1.0,
-                    pan: 0.0,
-                }],
-                length: 2.0,
-            },
-        ];
-
-        assert_eq!(compose.apply(vec_event1()), expected);
-    }
-
-    #[test]
-    fn op_sequence_test() {
-        let sequence_length_1 = sequence1().get_length_ratio();
-        let sequence_length_2 = sequence2().get_length_ratio();
-        assert_eq!(sequence_length_1, Ratio::new(5, 1));
-        assert_eq!(sequence_length_2, Ratio::new(10, 1));
-
-        let apply_expected = vec![
-            Event {
-                sounds: vec![Sound {
-                    frequency: 100.0,
-                    gain: 1.0,
-                    pan: 0.0,
-                }],
-                length: 1.0,
-            },
-            Event {
-                sounds: vec![Sound {
-                    frequency: 100.0,
-                    gain: 1.0,
-                    pan: 0.0,
-                }],
-                length: 1.0,
-            },
-            Event {
-                sounds: vec![Sound {
-                    frequency: 200.0,
-                    gain: 1.0,
-                    pan: 0.0,
-                }],
-                length: 1.0,
-            },
-            Event {
-                sounds: vec![Sound {
-                    frequency: 100.0,
-                    gain: 1.0,
-                    pan: 0.0,
-                }],
-                length: 2.0,
-            },
-        ];
-        assert_eq!(sequence1().apply(vec![event1()]), apply_expected);
-    }
-
-    #[test]
-    fn op_overlay_test() {
-        let overlay = Op::Overlay {
-            operations: vec![sequence1(), sequence2()],
-        };
-        assert_eq!(overlay.get_length_ratio(), Ratio::new(10, 1));
-
-        let apply_expected = vec![
-            Event {
-                sounds: vec![
-                    Sound {
-                        frequency: 100.0,
-                        gain: 1.0,
-                        pan: 0.0,
-                    },
-                    Sound {
-                        frequency: 100.0,
-                        gain: 1.0,
-                        pan: 0.0,
-                    },
-                ],
-                length: 1.0,
-            },
-            Event {
-                sounds: vec![
-                    Sound {
-                        frequency: 100.0,
-                        gain: 1.0,
-                        pan: 0.0,
-                    },
-                    Sound {
-                        frequency: 100.0,
-                        gain: 1.0,
-                        pan: 0.0,
-                    },
-                ],
-                length: 1.0,
-            },
-            Event {
-                sounds: vec![
-                    Sound {
-                        frequency: 200.0,
-                        gain: 1.0,
-                        pan: 0.0,
-                    },
-                    Sound {
-                        frequency: 100.0,
-                        gain: 1.0,
-                        pan: 0.0,
-                    },
-                ],
-                length: 1.0,
-            },
-            Event {
-                sounds: vec![
-                    Sound {
-                        frequency: 100.0,
-                        gain: 1.0,
-                        pan: 0.0,
-                    },
-                    Sound {
-                        frequency: 100.0,
-                        gain: 1.0,
-                        pan: 0.0,
-                    },
-                ],
-                length: 1.0,
-            },
-            Event {
-                sounds: vec![
-                    Sound {
-                        frequency: 100.0,
-                        gain: 1.0,
-                        pan: 0.0,
-                    },
-                    Sound {
-                        frequency: 200.0,
-                        gain: 1.0,
-                        pan: 0.0,
-                    },
-                ],
-                length: 1.0,
-            },
-            Event {
-                sounds: vec![Sound {
-                    frequency: 200.0,
-                    gain: 1.0,
-                    pan: 0.0,
-                }],
-                length: 1.0,
-            },
-            Event {
-                sounds: vec![Sound {
-                    frequency: 100.0,
-                    gain: 1.0,
-                    pan: 0.0,
-                }],
-                length: 4.0,
-            },
-        ];
-
-        assert_eq!(overlay.apply(vec![event1()]), apply_expected);
-    }
-}
