+++ src/write.rs
@@ -62,7 +62,11 @@ pub fn write_composition_to_json(
     let serialized = serde_json::to_string(&composition).unwrap();
     let mut file = File::create(format!("{}{}", filename, ".json".to_string()))?;
 
-    println!("{}.json was written and has {} event stream(s).", filename, composition.len());
+    println!(
+        "{}.json was written and has {} event stream(s).",
+        filename,
+        composition.len()
+    );
 
     file.write_all(serialized.as_bytes())?;
     Ok(())
